added new tests
