+++ test/Altinn.Platform.Events.Tests/TestingServices/OutboundServiceTests.cs
@@ -1,7 +1,11 @@
 using System;
+using System.Collections.Generic;
+using System.Threading;
+
 using Altinn.Platform.Events.Clients.Interfaces;
 using Altinn.Platform.Events.Configuration;
 using Altinn.Platform.Events.Models;
+using Altinn.Platform.Events.Repository;
 using Altinn.Platform.Events.Services;
 using Altinn.Platform.Events.Services.Interfaces;
 using Altinn.Platform.Events.Tests.Mocks;
@@ -23,6 +27,61 @@ namespace Altinn.Platform.Events.Tests.TestingServices
     /// </summary>
     public class OutboundServiceTests
     {
+        /// <summary>
+        /// Scenario: Event generated by an Altinn app is to be pushed
+        /// Expected result: Event source is modified to match sourceFilter in DB
+        /// Success criteria: Event source input to repository method has correct format
+        /// </summary>
+        [Fact]
+        public async void PostOutbound_AppSource_CorrectlyModifiedForRepositoryInput()
+        {
+            // Arrange
+            string expectedSimplified = "https://ttd.apps.altinn.no/ttd/endring-av-navn-v2";
+
+            CloudEvent cloudEvent = GetCloudEvent(new Uri("https://ttd.apps.altinn.no/ttd/endring-av-navn-v2/instances/1337/123124"), "/party/1337/", "app.instance.process.completed");
+
+            Mock<ISubscriptionRepository> repositoryMock = new();
+            repositoryMock
+                .Setup(r => r.GetSubscriptions(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<string>(), It.IsAny<CancellationToken>()))
+                .ReturnsAsync(new List<Subscription>());
+
+            var service = GetOutboundService(repositoryMock: repositoryMock.Object);
+
+            // Act
+            await service.PostOutbound(cloudEvent);
+
+            // Assert
+            repositoryMock.Verify(r => r.GetSubscriptions(It.Is<string>(s => s.Equals(expectedSimplified)), It.IsAny<string>(), It.IsAny<string>(), It.IsAny<CancellationToken>()));
+        }
+
+        /// <summary>
+        /// Scenario: Event generated by an external source is to be pushed
+        /// Expected result: Event source is not modified
+        /// Success criteria: Event source input to repository method matches event.source
+        /// </summary>
+        [Fact]
+        public async void PostOutbound_ExternalSource_UseEventSourceAsRepositoryInput()
+        {
+            // Arrange
+            string expectedSimplified = "urn:testing-events:test-source";
+           
+            CloudEvent cloudEvent = GetCloudEvent(new Uri("urn:testing-events:test-source"), "/party/1337/", "app.instance.process.completed");
+
+            Mock<ISubscriptionRepository> repositoryMock = new();
+            repositoryMock
+                .Setup(r => r.GetSubscriptions(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<string>(), It.IsAny<CancellationToken>()))
+                .ReturnsAsync(new List<Subscription>());
+
+            var service = GetOutboundService(repositoryMock: repositoryMock.Object);
+
+            // Act
+            await service.PostOutbound(cloudEvent);
+
+            // Assert
+            repositoryMock.Verify(r => r.GetSubscriptions(It.Is<string>(s => s.Equals(expectedSimplified)), It.IsAny<string>(), It.IsAny<string>(), It.IsAny<CancellationToken>()));
+        }
+
+
         /// <summary>
         /// Scenario:
         ///   Push an event. Three subscriptions are matching and is authorized
@@ -110,7 +169,76 @@ namespace Altinn.Platform.Events.Tests.TestingServices
             queueMock.VerifyAll();
         }
 
-        private static IOutboundService GetOutboundService(IEventsQueueClient queueMock = null, ILogger<IOutboundService> loggerMock = null)
+        /*
+             [Fact]
+        public async Task GetOrgSubscriptions_Two_Match()
+        {
+            SubscriptionService subscriptionService = GetSubscriptionService();
+            List<Subscription> result = await subscriptionService.GetSubscriptions(
+                "https://ttd.apps.altinn.no/ttd/endring-av-navn-v2",
+                "/party/1337",
+                "app.instance.process.completed");
+            Assert.True(result.Count == 2);
+        }
+
+        [Fact]
+        public async Task GetOrgSubscriptions_Zero_Match()
+        {
+            SubscriptionService subscriptionService = GetSubscriptionService();
+            List<Subscription> result = await subscriptionService.GetSubscriptions(
+                "https://ttd.apps.altinn.no/ttd/endring-av-navn-v1",
+                "/party/1337",
+                null);
+            Assert.True(result.Count == 0);
+        }
+
+        [Fact]
+        public async Task GetSubscriptions_One_Match()
+        {
+            SubscriptionService subscriptionService = GetSubscriptionService();
+            List<Subscription> result = await subscriptionService.GetSubscriptions(
+                "https://ttd.apps.altinn.no/ttd/new-app",
+                "/party/1337",
+                null);
+            Assert.True(result.Count == 1);
+        }
+
+        [Fact]
+        public async Task GetSubscriptions_Zero_Match()
+        {
+            SubscriptionService subscriptionService = GetSubscriptionService();
+            List<Subscription> result = await subscriptionService.GetSubscriptions(
+                "https://ttd.apps.altinn.no/ttd/endring-av-navn-v1",
+                "/party/1337",
+                null);
+            Assert.True(result.Count == 0);
+        }
+
+        [Fact]
+        public async Task GetOrgSubscriptions_SendsIncludeInvalidFalseToRepository()
+        {
+            _repositoryMock.Setup(rm => rm.GetSubscriptionsByConsumer(It.IsAny<string>(), false))
+                .ReturnsAsync(new List<Subscription>());
+
+            SubscriptionService subscriptionService = GetSubscriptionService(repository: _repositoryMock.Object);
+
+            await subscriptionService.GetSubscriptions("source", "party/1337", null);
+
+            _repositoryMock.VerifyAll();
+        }
+
+        [Fact]
+        public async Task GetOrgSubscriptions_InvalidSourceUri()
+        {
+            SubscriptionService subscriptionService = GetSubscriptionService();
+            List<Subscription> result = await subscriptionService.GetSubscriptions(
+                "ttd/endring-av-navn-v2",
+                "/party/1337",
+                "app.instance.process.completed");
+            Assert.True(result.Count == 0);
+        }
+         */
+        private static IOutboundService GetOutboundService(IEventsQueueClient queueMock = null, ISubscriptionRepository repositoryMock = null, ILogger<IOutboundService> loggerMock = null)
         {
             var services = new ServiceCollection();
             services.AddMemoryCache();
@@ -122,10 +250,20 @@ namespace Altinn.Platform.Events.Tests.TestingServices
                 loggerMock = new Mock<ILogger<IOutboundService>>().Object;
             }
 
+            if (queueMock == null)
+            {
+                queueMock = new EventsQueueClientMock();
+            }
+
+            if (repositoryMock == null)
+            {
+                repositoryMock = new SubscriptionRepositoryMock();
+            }
+
             IAuthorization authorizationMock = new AuthorizationService(new PepWithPDPAuthorizationMockSI());
             var service = new OutboundService(
                 queueMock,
-                new SubscriptionRepositoryMock(),
+               repositoryMock,
                 authorizationMock,
                 Options.Create(new PlatformSettings
                 {
