+++ test/Altinn.Platform.Events.Tests/TestingServices/SubscriptionServiceTest.cs
@@ -19,75 +19,7 @@ namespace Tests.TestingServices
     /// </summary>
     public class SubscriptionServiceTest
     {
-        private readonly Mock<ISubscriptionRepository> _repositoryMock = new();
-/*
-        [Fact]
-        public async Task GetOrgSubscriptions_Two_Match()
-        {
-            SubscriptionService subscriptionService = GetSubscriptionService();
-            List<Subscription> result = await subscriptionService.GetSubscriptions(
-                "https://ttd.apps.altinn.no/ttd/endring-av-navn-v2",
-                "/party/1337",
-                "app.instance.process.completed");
-            Assert.True(result.Count == 2);
-        }
-
-        [Fact]
-        public async Task GetOrgSubscriptions_Zero_Match()
-        {
-            SubscriptionService subscriptionService = GetSubscriptionService();
-            List<Subscription> result = await subscriptionService.GetSubscriptions(
-                "https://ttd.apps.altinn.no/ttd/endring-av-navn-v1",
-                "/party/1337",
-                null);
-            Assert.True(result.Count == 0);
-        }
-
-        [Fact]
-        public async Task GetSubscriptions_One_Match()
-        {
-            SubscriptionService subscriptionService = GetSubscriptionService();
-            List<Subscription> result = await subscriptionService.GetSubscriptions(
-                "https://ttd.apps.altinn.no/ttd/new-app",
-                "/party/1337",
-                null);
-            Assert.True(result.Count == 1);
-        }
-
-        [Fact]
-        public async Task GetSubscriptions_Zero_Match()
-        {
-            SubscriptionService subscriptionService = GetSubscriptionService();
-            List<Subscription> result = await subscriptionService.GetSubscriptions(
-                "https://ttd.apps.altinn.no/ttd/endring-av-navn-v1",
-                "/party/1337",
-                null);
-            Assert.True(result.Count == 0);
-        }
-
-        [Fact]
-        public async Task GetOrgSubscriptions_SendsIncludeInvalidFalseToRepository()
-        {
-            _repositoryMock.Setup(rm => rm.GetSubscriptionsByConsumer(It.IsAny<string>(), false))
-                .ReturnsAsync(new List<Subscription>());
-
-            SubscriptionService subscriptionService = GetSubscriptionService(repository: _repositoryMock.Object);
-
-            await subscriptionService.GetSubscriptions("source", "party/1337", null);
-
-            _repositoryMock.VerifyAll();
-        }
-
-        [Fact]
-        public async Task GetOrgSubscriptions_InvalidSourceUri()
-        {
-            SubscriptionService subscriptionService = GetSubscriptionService();
-            List<Subscription> result = await subscriptionService.GetSubscriptions(
-                "ttd/endring-av-navn-v2",
-                "/party/1337",
-                "app.instance.process.completed");
-            Assert.True(result.Count == 0);
-        }*/
+        private readonly Mock<ISubscriptionRepository> _repositoryMock = new();    
 
         [Fact]
         public async Task GetAllSubscriptions_SendsIncludeInvalidTrueToRepository()
